refactored driver: removed wrapper and split real and gazebo drivers
