--- src/drivers/PiBot/__init__.py
@@ -1,1 +1,0 @@
-from .PiBot import dameRobot
