--- src/drivers/PiBot/gazebo/__init__.py
@@ -1,1 +1,0 @@
-from .piBot import PiBot
