--- src/drivers/PiBot/real/__init__.py
@@ -1,1 +1,0 @@
-from .piBot import PiBot
