+++ src/drivers/PiBot_gz/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Install resources
+INSTALL(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/gazebo DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python2.7/PiBot COMPONENT setup)
+INSTALL(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/real DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python2.7/PiBot COMPONENT setup)
+
+INSTALL (FILES ${CMAKE_CURRENT_SOURCE_DIR}/Kibotics.yml DESTINATION ${CMAKE_INSTALL_PREFIX}/share/jderobot/conf COMPONENT pibot)
+
+INSTALL (FILES ${CMAKE_CURRENT_SOURCE_DIR}/PiBot.py DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python2.7/PiBot COMPONENT pibot)
+INSTALL (FILES ${CMAKE_CURRENT_SOURCE_DIR}/__init__.py DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python2.7/PiBot COMPONENT pibot)
